fix: Correct all remaining API mismatches
- Fix HNSW method: HnswSetEfSearch → SetEfSearch
- Fix NewIndexIVFPQ: remove metric parameter (5th arg)
- Fix ScalarQuantizer: ScalarQuantizerType_QT_8bit → QT_8bit
- Fix Ntotal() comparison: convert int to int64
- Remove unused 'time' import from hnsw_recall_test.go
- Fix buildHNSW functions to capture errors

All test packages now parse successfully.

+++ test/recall/framework.go
@@ -166,7 +166,7 @@ func RunRecallTest(t *testing.T, config RecallTestConfig) RecallTestResult {
 		return result
 	}
 
-	if index.Ntotal() != n {
+	if index.Ntotal() != int64(n) {
 		result.Error = fmt.Errorf("expected %d vectors in index, got %d", n, index.Ntotal())
 		t.Error(result.Error)
 		return result
